Update README.md
The files to reconfigure the standard VNAO ready room into the VFA 469 livery (as used on our server).  Please note that most of the files will be static, but if you want to see updates to the greenie board you should allow the repo to push changes to your local repo.